+++ crates/rvr/rvr-openvm/src/emit/project.rs
@@ -105,7 +105,7 @@ pub struct CProject {
     /// `None` in pure mode (no chip metadata requested); must be set in metered modes.
     pub pc_to_chip: Option<Vec<TraceChipIndex>>,
     /// Program PC base (used to compute pc_to_chip index).
-    pub pc_base: u32,
+    pub pc_base: u64,
     /// Per-AIR widths for MeteredCost precomputation. Indexed by chip index.
     pub chip_widths: Option<Vec<u64>>,
     /// Compile with native debug info (`-g -fno-omit-frame-pointer`).
@@ -318,11 +318,11 @@ impl CProject {
             .replace('\t', "\\t")
     }
 
-    fn block_end_pc(&self, block: &Block) -> u32 {
+    fn block_end_pc(&self, block: &Block) -> u64 {
         block.terminator_pc.saturating_add(4)
     }
 
-    fn dispatch_max_pc(blocks: &[Block], entry_point: u32, text_start: u32) -> u32 {
+    fn dispatch_max_pc(blocks: &[Block], entry_point: u64, text_start: u64) -> u64 {
         blocks
             .iter()
             .map(|b| b.start_pc)
@@ -331,7 +331,7 @@ impl CProject {
             .unwrap_or(text_start)
     }
 
-    fn dispatch_table_size(text_start: u32, text_end: u32) -> usize {
+    fn dispatch_table_size(text_start: u64, text_end: u64) -> usize {
         debug_assert!(text_end >= text_start);
         ((text_end - text_start) / 4 + 1) as usize
     }
@@ -358,7 +358,7 @@ impl CProject {
 
     /// Look up the chip index for a given PC. Must only be called in metered
     /// modes; panics if `pc_to_chip` is unset.
-    fn chip_idx_for_pc(&self, pc: u32) -> TraceChipIndex {
+    fn chip_idx_for_pc(&self, pc: u64) -> TraceChipIndex {
         let mapping = self
             .pc_to_chip
             .as_ref()
@@ -376,8 +376,8 @@ impl CProject {
     pub fn write_all<F: PrimeField32>(
         &self,
         blocks: &[Block],
-        entry_point: u32,
-        text_start: u32,
+        entry_point: u64,
+        text_start: u64,
         extensions: &ExtensionRegistry<F>,
     ) -> io::Result<()> {
         let text_end = Self::dispatch_max_pc(blocks, entry_point, text_start);
@@ -398,8 +398,8 @@ impl CProject {
 
     fn write_constants(
         &self,
-        text_start: u32,
-        text_end: u32,
+        text_start: u64,
+        text_end: u64,
         dispatch_table_size: usize,
     ) -> io::Result<()> {
         let h = constants_header(text_start, text_end, dispatch_table_size);
@@ -597,7 +597,7 @@ impl CProject {
         let num_partitions = blocks.len().div_ceil(self.blocks_per_partition);
 
         // Precompute valid block PCs for tail-call target validation.
-        let valid_blocks: HashSet<u32> = blocks.iter().map(|b| b.start_pc).collect();
+        let valid_blocks: HashSet<u64> = blocks.iter().map(|b| b.start_pc).collect();
 
         for part_idx in 0..num_partitions {
             let start = part_idx * self.blocks_per_partition;
@@ -621,7 +621,7 @@ impl CProject {
         Ok(())
     }
 
-    fn emit_block_function(&self, out: &mut String, block: &Block, valid_blocks: &HashSet<u32>) {
+    fn emit_block_function(&self, out: &mut String, block: &Block, valid_blocks: &HashSet<u64>) {
         let pc = block.start_pc;
         let end_pc = self.block_end_pc(block);
         let insn_count = block.insn_count();
@@ -727,7 +727,7 @@ impl CProject {
         writeln!(out).unwrap();
     }
 
-    fn emit_segment_checkpoint(&self, out: &mut String, pc: u32) {
+    fn emit_segment_checkpoint(&self, out: &mut String, pc: u64) {
         writeln!(
             out,
             "    MeteredSegmentCheckpointResult checkpoint = metered_segment_checkpoint(state, check_counter);"
@@ -758,7 +758,7 @@ impl CProject {
         self.emit_instret_suspend_check(out, pc, insn_count);
     }
 
-    fn emit_metered_counter_check(&self, out: &mut String, pc: u32, insn_count: u32) {
+    fn emit_metered_counter_check(&self, out: &mut String, pc: u64, insn_count: u32) {
         let args = self.fn_args_from_params();
         writeln!(out, "    if (unlikely(check_counter < {insn_count}u)) {{").unwrap();
         writeln!(
@@ -769,7 +769,7 @@ impl CProject {
         writeln!(out, "    }}").unwrap();
     }
 
-    fn emit_instret_suspend_check(&self, out: &mut String, pc: u32, insn_count: u32) {
+    fn emit_instret_suspend_check(&self, out: &mut String, pc: u64, insn_count: u32) {
         debug_assert_ne!(
             self.tracer_mode,
             TracerMode::Metered,
@@ -784,7 +784,7 @@ impl CProject {
         writeln!(out, "    }}").unwrap();
     }
 
-    fn emit_suspend_return(&self, out: &mut String, pc: u32) {
+    fn emit_suspend_return(&self, out: &mut String, pc: u64) {
         let save = self.save_hot_regs_call();
         writeln!(out, "        {save}").unwrap();
         writeln!(
@@ -811,7 +811,7 @@ impl CProject {
         }
 
         let mut chip_counts: BTreeMap<u32, u32> = BTreeMap::new();
-        let mut increment_chip_count = |pc: u32| match self.chip_idx_for_pc(pc) {
+        let mut increment_chip_count = |pc: u64| match self.chip_idx_for_pc(pc) {
             TraceChipIndex::Chip(chip) => *chip_counts.entry(chip.as_u32()).or_insert(0) += 1,
             TraceChipIndex::NoChip => {}
         };
@@ -851,7 +851,7 @@ impl CProject {
     fn emit_source_annotation(
         &self,
         out: &mut String,
-        pc: u32,
+        pc: u64,
         opname: &str,
         source_loc: Option<&SourceLoc>,
     ) {
@@ -876,8 +876,8 @@ impl CProject {
     fn write_dispatch(
         &self,
         blocks: &[Block],
-        entry_point: u32,
-        text_start: u32,
+        entry_point: u64,
+        text_start: u64,
     ) -> io::Result<()> {
         let name = &self.name;
         let mut src = String::with_capacity(16 * 1024);
@@ -914,12 +914,12 @@ impl CProject {
         let table_size = Self::dispatch_table_size(text_start, max_pc);
 
         // Build block-start lookup.
-        let block_starts: std::collections::HashMap<u32, u32> =
+        let block_starts: std::collections::HashMap<u64, u64> =
             blocks.iter().map(|b| (b.start_pc, b.start_pc)).collect();
 
         writeln!(src, "BlockFn dispatch_table[RV_DISPATCH_TABLE_SIZE] = {{").unwrap();
         for i in 0..table_size {
-            let pc = text_start + (i as u32) * 4;
+            let pc = text_start + (i as u64) * 4;
             if block_starts.contains_key(&pc) {
                 writeln!(src, "    block_0x{pc:08x},").unwrap();
             } else {
